Use two_electron keyword instead of filling
Previously, the switch between the 4-level and 6-level model with
two-electron states was made with the filling keyword such that
filling=1 and filling=3 had the exact same behavior. This was not
transparent to the user, especially since the get_ex_spectrums
had DIFFERENT behaviors for filling=1 and filling=3

Now, the switch between the 4- and 6-level models is made with
the two_electron=True/False keyword. get_ex_spectrums still
has the filling keyword since for that function it actually
results in different behavior.